pyproject.toml
[project]
name = "pcmap"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "ccmap>=4.1",
    "numpy>=2.0.2",
    "pypstruct>=0.1.1",
    "rich-cli>=1.8.1",
    "rich-click>=1.9.7",
]

[project.scripts]
pcmap-monomer = "pcmap.bin.monomer:compute_monomer"
pcmap-dimer   = "pcmap.bin.dimer:compute_dimer"
pcmap-zdock   = "pcmap.bin.z_dimer:compute_z_dimer"
pcmap-many   = "pcmap.bin.many:compute_many"

[build-system]
requires = ["uv_build>=0.10.0,<0.11.0"]
build-backend = "uv_build"
